move .xs tests to test_xs.py
